aux-ml/requirements.txt
fastapi==0.115.13
httpx==0.28.1
pydantic==2.11.4
PyYAML==6.0.2
pymupdf==1.27.2
uvicorn==0.34.2
